+++ src/Messenger/Stamp/AzureMessageStamp.php
@@ -4,31 +4,42 @@ declare(strict_types=1);
 
 namespace AymDev\MessengerAzureBundle\Messenger\Stamp;
 
-use Symfony\Component\Messenger\Stamp\NonSendableStampInterface;
+use Symfony\Component\Messenger\Stamp\StampInterface;
 use Symfony\Contracts\HttpClient\Exception\HttpExceptionInterface;
 use Symfony\Contracts\HttpClient\Exception\TransportExceptionInterface;
 use Symfony\Contracts\HttpClient\ResponseInterface;
 
 /**
- * Azure Service Bus receiver stamp with message metadata
- * TODO: remove stamp in v2
- * @deprecated AzureReceivedStamp is kept for compatibility purposes. Use AzureMessageStamp instead
+ * Azure Service Bus stamp added to sent and received messages containing message metadata
  */
-class AzureReceivedStamp implements NonSendableStampInterface
+class AzureMessageStamp implements StampInterface
 {
     /** @var string */
+    private $entityPath;
+
+    /** @var string */
     private $message;
+
+    /** @var null|string */
+    private $subscriptionName;
 
     /** @var string|null */
     private $locationHeader;
 
     /**
+     * @internal
      * @param string $message the original message
      * @param string|null $locationHeader optional "Location" response header
      */
-    public function __construct(string $message, ?string $locationHeader)
-    {
+    public function __construct(
+        string $entityPath,
+        string $message,
+        ?string $subscriptionName = null,
+        ?string $locationHeader = null
+    ) {
+        $this->entityPath = $entityPath;
         $this->message = $message;
+        $this->subscriptionName = $subscriptionName;
         $this->locationHeader = $locationHeader;
     }
 
@@ -37,17 +48,32 @@ class AzureReceivedStamp implements NonSendableStampInterface
      * @internal
      * @throws HttpExceptionInterface|TransportExceptionInterface
      */
-    public static function createFromResponse(ResponseInterface $response): self
-    {
+    public static function createFromResponse(
+        ResponseInterface $response,
+        string $entityPath,
+        ?string $subscriptionName
+    ): self {
         return new self(
+            $entityPath,
             $response->getContent(),
+            $subscriptionName,
             $response->getHeaders()['location'][0] ?? null
         );
+    }
+
+    public function getEntityPath(): string
+    {
+        return $this->entityPath;
     }
 
     public function getMessage(): string
     {
         return $this->message;
+    }
+
+    public function getSubscriptionName(): ?string
+    {
+        return $this->subscriptionName;
     }
 
     public function getLocationHeader(): ?string